// src/app/shared/types/api-types/bido-notification-dto.interface.ts
import { BidoEquipmentDTO } from './bido-equipment-dto.interface';
import { BidoNotificationTypeDTO } from './bido-notification-type-dto.interface';
import { BidoRootCauseDTO } from './bido-root-cause-dto.interface';

export interface BidoNotificationDTO {
  notificationCode?: string;
  notNotificationCode?: string;
  typeCode?: string;
  rootCauseId?: string;
  notificationSubject?: string;
  notificationDescription?: string;
  reportingPeriodStartDate?: Date;
  reportingPeriodStartDateStr?: string;
  reportingPeriodEndDate?: Date;
  equipmentCode?: string;
  status?: string;
  priority?: string;
  isScheduled?: boolean;
  eventCause?: string;
  notificationRemarks?: string;
  taskCode?: string;
  taskVersion?: string;
  familyVariantCode?: string;
  fieldName?: string;
  fieldNewValue?: string;
  fieldOldValue?: string;
  chapter?: string;
  section?: string;
  subject?: string;
  sheet?: string;
  marks?: string;
  position?: string;
  ownerCode?: string;
  operatorCode?: string;
  zone?: string;
  country?: string;
  eventReason?: string;
  isDeduced?: boolean;
  topAircraftCode?: string;
  topEngineCode?: string;
  reportingDate?: Date;
  reportingDateStr?: String;
  movementPn?: string;
  movementQuantity?: number;
  taskAssetPn?: string;
  taskAssetCode?: string;
  taskAtaChapter?: string;
  taskAtaSection?: string;
  taskAtaSubject?: string;
  hasFmd?: boolean;
  eventSequenceCode?: string;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
  bidoRootCauseDTO?: BidoRootCauseDTO;
  bidoNotificationTypeDTO: BidoNotificationTypeDTO;
  bidoEquipmentDTO?: BidoEquipmentDTO;
}